refactor: Move impls near definitions
